refactor(header): extract ThemeSwitch to a separate file
Moved the ThemeSwitch component from header.tsx to a new file,
theme-switch.tsx. This change improves code organization and
readability by isolating the theme-switch logic. Adjusted imports
accordingly to reflect this separation.
